+++ CHANGELOG.md
@@ -8,6 +8,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - `_classname` property has been completely removed, officially dropping support for Neo4j < 2.1.5.
 
+### Added
+
+- Optional three-argument signature for `ActiveRel#create` and `#create!`, just like `initialize`.
+
 ## [6.0.0.alpha.11] - 11-3-2015
 
 ### Fixed